Add base directory as arguments to the upload functions in cvd_utils
To support multiple cuttlefish sharing one user account on a remote
host, each instance's files will be moved from HOME to a different base
directory. The functions that access the files need the directory path
as an argument. This commit adds arguments to the utility functions that
upload cuttlefish images. This does not change the file locations.

Test: acloud-dev create --avd-type cuttlefish \
      --host 192.168.9.2 --host-ssh-private-key-path ~/id_rsa \
      --local-image ~/img --local-kernel-image ~/img \
      --cvd-host-package ~/cvd-host_package.tar.gz \
      --report-file ~/report.json
Test: acloud-dev create --avd-type cuttlefish \
      --local-image ~/img --local-kernel-image ~/img \
      --cvd-host-package ~/cvd-host_package.tar.gz \
      --report-file ~/report.json
Bug: 229812494
Change-Id: I7184d0b5f97ad47c2c3f4cb5d7668e945c3c03e8