Upload local emulator zip to remote host
In the continuous integration infrastructure, acloud processes may
concurrently download goldfish artifacts and corrupt the
--image-download-dir. The synchronzation between acloud processes is
complicated. A simpler method to avoid corrupting the files is to let
the infrastructure download the files and invoke acloud with local
files.

This commit adds --emulator-zip. The option is followed by the emulator
zip path that is going to be uploaded to the remote host.

Test: acloud-dev create --avd-type goldfish -vv \
      --build-id 9714076 \
      --build-target sdk_slim_x86_64-userdebug \
      --emulator-zip ~/sdk-repo-linux-emulator-9714371.zip \
      --host 192.168.9.2 --host-ssh-private-key-path ~/id_rsa
Bug: 245226952
Change-Id: I602e56bb358279604e8605338076fb4fba86603e