acloud: Upload host package directory if available, instead of the tarball.
A proposed change to the build system will cause it to no longer create
the host package tarball when issuing the `m` command. Instead, the
contents of the host package are installed to a specific directory. This
change teaches acloud to upload the contents of the directory if it
exists, in preference to the tarball, so that the build system change
will not break acloud create --local-image.

Bug: 273862156
Change-Id: Ib788190e53040a1156c46795546802521735c3c5