Use `ssh -t` on `launch_cvd` to acommodate the `cvd` server version
When the initial `launch_cvd` call starts the `cvd` server, it hangs on
to a `ssh` pipe, causing `ssh` to never terminate. This accommodates the
behavior of `cvd` by using the pseudo-tty option of ssh, which also
changes its behavior to wait for the process to exit rather than the
pipes to close.

Bug: 293191537
Test: Apply aosp/2713515. m acloud-dev && acloud-dev create --local-image -vv
Change-Id: I5d52451441177941d82dc5bd1812d8e84b7bb54e