Use raw string to wrap the welcome message
In my own setup, I keep seeing things like
```
/tmp/Soong.python_w5uotf2f/acloud/setup/setup.py:89: SyntaxWarning: invalid escape sequence '\/'
  "  / _ |/ ___/ /  / __ \/ / / / _ \\ \n" \
```

This is because the existence of "\/" in the welcome message string.

This CL wrap the welcome message with the raw string to fix the issue.
Additionally, this helps people to read the message in the code easier
due to the removal of escape characters.

Bug: 399021454
Test: execute `acloud-dev setup` and not seeing any error message
Change-Id: I2928ef443a0378a4e51f6a7553603e713b2f79a0