Let _SshLogOutput raise TimeoutExpired
_SshLogOutput calls Popen and kills it when timeout. Without this patch,
the Popen object returns non-zero, and the return code is converted to
CalledProcessError. The caller cannot determine whether the error is the
result of the command or a timeout.

This patch makes _SshLogOutput raise TimeoutExpired. All of the callers
that catch CalledProcessError should catch TimeoutExpired as well. This
patch also removes unused _SshCall and _SshCallWait.

Test: atest acloud_test
Test: acloud-dev create --local-image --boot-timeout 10 -vv
Bug: 401549262
Change-Id: I722580ca726aaadd28bfff679e25bb58a773fa12