cuttlefish: Remove fetch_cvd usage from acloud
fetch_cvd was deprecated and deleted in 2023.
The last readily available build of fetch cvd was recently
purged and all acloud creates now fail due to its absence.

To mitigate this - migrate all fetching to use the cvd
binary. cvd is part of the Cloud Android Host image
so it no longer needs to be fetched.

Bug: 402627490
Test: m acloud; acloud-dev create
Change-Id: I1ad4099888971fe071c7a8315eac811893815eb2