Add the fetch.log to the local report logs list
To match the remote instance cases.

The remote image, local instance case falls back to the local image,
local instance case after a fetch.  That means the logic needs to be
able to handle both cases, where the file may or may not have been created.
The `image_dir` is used for `--target_directory`, which is where `cvd fetch`
will put the log file.

Test: m acloud
Test: acloud-dev create --local-instance --branch=git_main
--build-target=aosp_cf_x86_64_only_phone-trunk_staging-userdebug
--report-file=report.json
Test: cat report.json | grep fetch.log

Change-Id: If6478052448ba293091a8d056470ba96ca3b3129